+++ RELEASE_NOTES.md
@@ -1,3 +1,5 @@
+* 0.7.3 - Add GZipStream compression implementation.
+* 0.7.2 - Revert changes to AppDomain pool.
 * 0.7.1 - Expose data compression abstractions; Tie AppDomain pool API to Vagabond.
 * 0.7.0 - Refinements to the API, changes in the Assembly loading logic.
 * 0.6.12 - Changes to assembly loading logic.